gradient-infrastructure/requirements.txt
invoke==1.6.0
attrs==21.2.0
boto3==1.18.26
grpcio-tools==1.33.2
twine==3.4.2
aws-cdk.core==1.123.0
aws-cdk.aws-s3==1.123.0
aws-cdk.aws-codebuild==1.123.0
aws-cdk.aws-codepipeline==1.123.0
aws-cdk.aws-ecr==1.123.0
aws-cdk.aws-codeartifact==1.123.0
aws-cdk.aws-iam==1.123.0
pytest==6.2.5
pyaml==21.8.3